pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.69.0",
    "flask>=3.1.2",
    "gitpython>=3.1.45",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
